+++ plugins/domains/mashable.com.js
@@ -45,8 +45,9 @@ module.exports = {
             html += $html.html();
 
             return {
-                html_for_readability: html,
-                ignore_readability_error: true
+                readability_data: {
+                    html: html
+                }
             };
         }
     },