+++ WHATSNEW.md
@@ -1,3 +1,15 @@
+2013.08.13, Version 0.4.4
+
+ + added tumblr.com meta plugin
+ + added tumblr.com photo plugin
+ + added tumblr.com video plugin
+ + added tumblr.com text plugin
+ * fixed animoto.com plugin
+ * updated guardian.com plguin domain
+ * updated businessinsider.com plguin
+ * fixed slideshare.net plugin
+ + added huffingtonpost.com plugin
+
 2013.08.08, Version 0.4.3
 
  + added mixbit.com plugin